extract action running part to object class
This lets us use this also in dynamically created actions which can't inherit from this.
